requirements.txt
openai==0.27.7
pinecone-client[grpc]==3.0.0
pinecone-datasets==0.7.0
tqdm
python-dotenv==1.0.0
